Remove unnecessary check for node[:mesos][:master]
As there is a default value for this attribute, this condition will always
be true, so there is no need to check for it (the edge case of a user
removing the attribute from a node after-the-fact is not something
we need to guard against).

Additionally, drop default values from the resources.

Less is more!