Drop default values from various resources
While explicit, having these results in much more difficult to read
recipes, and is frankly just noise. One may refer to the Chef docs to
know the default values, if need be.

W/r/t to the templates, they will take their source from the filename
that they create, minus the path.

As for notifications, they are, by default, `:delayed`.