Refactor & DRY up master/slave configs & services
Rather than have the config templates and services in the install-type
recipes, we can use node[:mesos][:type] to determine, in the correct
master/slave recipe in use, which template to populate, and which
service to enable/start/stop.

With the previous system, regardless of the install type, a node would
end up being both a master and slave (i.e., both configs would be
populated, and both services [re]started).

This avoids that behavior, promoting a separation of concerns.